requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.1.7
django-dotenv==1.4.2
idna==2.10
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
psycopg2==2.8.6
pycparser==2.20
pyOpenSSL==20.0.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
coreapi==2.3.3
coreschema==0.0.4
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
PyJWT==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
uritemplate==3.0.1

